drf-project/users/urls.py
from django.urls import path
from rest_framework_simplejwt.views import TokenRefreshView
from .views import SignupView, SigninView, MeView

urlpatterns = [
    path('signup/', SignupView.as_view(), name='signup'),
    path('signin/', SigninView.as_view(), name='signin'),
    path('me/', MeView.as_view(), name='me'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]